+++ modules/router/src/router.ts
@@ -50,62 +50,58 @@ export class Router implements IRouter {
   }
 
   private configureMetrics() {
-    // Track the total number of channels
-    const channelCounter = new Gauge({
-      name: "router_channels_total",
-      help: "router_channels_total_help",
-      registers: [this.register],
-    });
-
-    const collateral = new Gauge({
-      name: "router_channels_collateral",
-      help: "router_channels_collateral_help",
-      labelNames: ["assetId", "channelAddress"],
-      registers: [this.register],
-    });
-
-    // TODO: fix this once this issue is fixed by using the `collect` function in the gauge
-    // https://github.com/siimon/prom-client/issues/383
-    setInterval(async () => {
-      this.logger.debug({}, "Collecting metrics");
-      const channels = await this.nodeService.getStateChannels({ publicIdentifier: this.publicIdentifier });
-      if (channels.isError) {
-        this.logger.error(
-          { error: channels.getError()!.message, publicIdentifier: this.publicIdentifier },
-          "Failed to fetch channels",
-        );
-        return;
-      }
-      const channelAddresses = channels.getValue();
-      channelCounter.set(channelAddresses.length);
-
-      for (const channelAddr of channelAddresses) {
-        const channelState = await this.nodeService.getStateChannel({
-          channelAddress: channelAddr,
-          publicIdentifier: this.publicIdentifier,
-        });
-        if (channelState.isError) {
-          this.logger.error(
-            { error: channelState.getError()!.message, channelAddress: channelAddr },
-            "Failed to get channel",
-          );
-          return;
-        }
-        const { balances, assetIds, aliceIdentifier } = channelState.getValue() as FullChannelState;
-        assetIds.forEach((assetId: string, index: number) => {
-          const balance = balances[index];
-          if (!balance) {
-            return;
-          }
-          // Set the proper collateral gauge
-          collateral.set(
-            { assetId, channelAddress: channelAddr },
-            parseFloat(formatEther(balance.amount[this.publicIdentifier === aliceIdentifier ? 0 : 1])),
-          );
-        });
-      }
-
-      this.logger.debug({}, "Done collecting metrics");
-    }, 30_000);
+    // // Track the total number of channels
+    // const channelCounter = new Gauge({
+    //   name: "router_channels_total",
+    //   help: "router_channels_total_help",
+    //   registers: [this.register],
+    // });
+    // const collateral = new Gauge({
+    //   name: "router_channels_collateral",
+    //   help: "router_channels_collateral_help",
+    //   labelNames: ["assetId", "channelAddress"],
+    //   registers: [this.register],
+    // });
+    // // TODO: fix this once this issue is fixed by using the `collect` function in the gauge
+    // // https://github.com/siimon/prom-client/issues/383
+    // setInterval(async () => {
+    //   this.logger.debug({}, "Collecting metrics");
+    //   const channels = await this.nodeService.getStateChannels({ publicIdentifier: this.publicIdentifier });
+    //   if (channels.isError) {
+    //     this.logger.error(
+    //       { error: channels.getError()!.message, publicIdentifier: this.publicIdentifier },
+    //       "Failed to fetch channels",
+    //     );
+    //     return;
+    //   }
+    //   const channelAddresses = channels.getValue();
+    //   channelCounter.set(channelAddresses.length);
+    //   for (const channelAddr of channelAddresses) {
+    //     const channelState = await this.nodeService.getStateChannel({
+    //       channelAddress: channelAddr,
+    //       publicIdentifier: this.publicIdentifier,
+    //     });
+    //     if (channelState.isError) {
+    //       this.logger.error(
+    //         { error: channelState.getError()!.message, channelAddress: channelAddr },
+    //         "Failed to get channel",
+    //       );
+    //       return;
+    //     }
+    //     const { balances, assetIds, aliceIdentifier } = channelState.getValue() as FullChannelState;
+    //     assetIds.forEach((assetId: string, index: number) => {
+    //       const balance = balances[index];
+    //       if (!balance) {
+    //         return;
+    //       }
+    //       // Set the proper collateral gauge
+    //       collateral.set(
+    //         { assetId, channelAddress: channelAddr },
+    //         parseFloat(formatEther(balance.amount[this.publicIdentifier === aliceIdentifier ? 0 : 1])),
+    //       );
+    //     });
+    //   }
+    //   this.logger.debug({}, "Done collecting metrics");
+    // }, 30_000);
   }
 }