+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,12 @@
 
 ## Next Release
 
+## 0.1.0-rc.19
+
+- \[router\] Remove obtrusive metrics
+- \[messaging\] Properly catch messaging timeouts
+- \[ops\] Fix router logdna
+
 ## 0.1.0-rc.18
 
 - \[all\] Fix bad publishing